backend/django_recipe_api/requirements.txt
django
djongo
djangorestframework
django-cors-headers